pyproject.toml
[tool.poetry]
name = "google-cal"
version = "0.1.0"
description = ""
authors = ["Isabelle-C <isabellechen2023@u.northwestern.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-api-python-client = "^2.101.0"
google-auth-httplib2 = "^0.1.1"
google-auth-oauthlib = "^1.1.0"
pyyaml = "^6.0.1"
pytz = "^2023.3.post1"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
